ncal: fix size_t passed as the int * field width
mw - wcslen(...) - 1 promotes to size_t but the * directive consumes
an int — same varargs UB class (benign on LP64 LE, wrong on
big-endian). Cast to (int), matching the file's own wcenter/center
precedent. Verbatim upstream.

+++ src.freebsd/miscutils/ncal/ncal.c
@@ -743,7 +743,8 @@ monthranger(int y, int m, int jd_flag, int before, int after)
 				wprintf(L"%-*ls", mw, year[i].name);
 			else
 				wprintf(L"%-ls %-*d", year[i].name,
-				    mw - wcslen(year[i].name) - 1, M2Y(m + i));
+				    (int)(mw - wcslen(year[i].name) - 1),
+				    M2Y(m + i));
 		wprintf(L"\n");
 
 		/* And the days of the month. */
